Added option to give .5 ratings. And some seeds for reviews.

+++ database/migrations/2026_05_18_000001_create_ratings_table.php
@@ -12,7 +12,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnDelete();
             $table->foreignId('recipe_id')->constrained()->cascadeOnDelete();
-            $table->unsignedTinyInteger('value');
+            $table->decimal('value', 2, 1);
             $table->text('review')->nullable();
             $table->timestamps();
